Move trace extraction to common reusable utility
