+++ client/thin-replica-client/include/client/thin-replica-client/trace_contexts.hpp
@@ -15,6 +15,7 @@
 #define THIN_REPLICA_CLIENT_TRACE_CONTEXTS_HPP_
 
 #include <opentracing/span.h>
+#include <grpcpp/server.h>
 #include <optional>
 
 #include "thin_replica.pb.h"
@@ -30,12 +31,14 @@ namespace client::thin_replica_client {
 class TraceContexts {
  public:
   using SpanPtr = std::unique_ptr<opentracing::Span>;
+  using SpanCtxPtr = std::unique_ptr<opentracing::SpanContext>;
 
   static void InjectSpan(const SpanPtr& span, cc::EventVariant& update);
   static expected<std::unique_ptr<opentracing::SpanContext>> ExtractSpan(const cc::EventVariant& update);
   static SpanPtr CreateChildSpanFromBinary(const std::string& trace_context,
                                            const std::string& child_name,
                                            const std::string& correlation_id = {});
+  static SpanCtxPtr ExtractSpanFromMetadata(const opentracing::Tracer& tracer, const grpc::ServerContext& context);
 };
 
 }  // namespace client::thin_replica_client
