+++ client/thin-replica-client/src/trace_contexts.cpp
@@ -111,4 +111,47 @@ SpanPtr TraceContexts::CreateChildSpanFromBinary(const std::string& trace_contex
   return opentracing::Tracer::Global()->StartSpan(
       child_name, {opentracing::FollowsFrom(&**parent_span_context), opentracing::SetTag{kCorrelationIdTag, cid}});
 }
+
+struct GrpcMetadataCarrierForReading : opentracing::TextMapReader {
+  GrpcMetadataCarrierForReading(const std::multimap<grpc::string_ref, grpc::string_ref>& client_metadata)
+      : client_metadata_(client_metadata) {}
+
+  using F = std::function<opentracing::expected<void>(opentracing::string_view, opentracing::string_view)>;
+
+  opentracing::expected<void> ForeachKey(F f) const override {
+    // Iterate through all key-value pairs, the tracer will use the relevant keys
+    // to extract a span context.
+    for (auto& key_value : client_metadata_) {
+      auto was_successful = f(FromStringRef(key_value.first), FromStringRef(key_value.second));
+      if (!was_successful) {
+        // If the callback returns and unexpected value, bail out of the loop.
+        return was_successful;
+      }
+    }
+    // Indicate successful iteration.
+    return {};
+  }
+
+  opentracing::expected<opentracing::string_view> LookupKey(opentracing::string_view key) const override {
+    auto find_it = client_metadata_.find(grpc::string_ref(key));
+    if (find_it != client_metadata_.end()) {
+      return opentracing::make_unexpected(opentracing::key_not_found_error);
+    }
+    return opentracing::string_view{FromStringRef(find_it->second)};
+  }
+
+ private:
+  static std::string FromStringRef(const grpc::string_ref& string_ref) {
+    return std::string(string_ref.data(), string_ref.size());
+  }
+
+  const std::multimap<grpc::string_ref, grpc::string_ref>& client_metadata_;
+};
+
+TraceContexts::SpanCtxPtr TraceContexts::ExtractSpanFromMetadata(const opentracing::Tracer& tracer,
+                                                                 const grpc::ServerContext& context) {
+  GrpcMetadataCarrierForReading metadata_carrier(context.client_metadata());
+  return *tracer.Extract(metadata_carrier);
+}
+
 }  // namespace client::thin_replica_client
